# expression: implement vectorized evaluation for builtinMakeDateSig